Revert "Scalar Api documentation"
