+++ ExpenseTracker/Program.cs
@@ -1,9 +1,7 @@
-using Scalar.AspNetCore;
 using ExpenseTracker.Data;
-using ExpenseTracker.Services;
 using Microsoft.EntityFrameworkCore;
 
-WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
+var builder = WebApplication.CreateBuilder(args);
 
 // Add database connection
 builder.Services.AddDbContext<ExpenseTrackerDbContext>(options =>
@@ -11,27 +9,18 @@ builder.Services.AddDbContext<ExpenseTrackerDbContext>(options =>
         new MySqlServerVersion(new Version(8, 0, 21))));
 
 
-// Register application services
-builder.Services.AddOpenApi();
-builder.Services.AddScoped<IExpenseService, ExpenseService>();
-
 // Add services to the container
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
 
-WebApplication app = builder.Build();
+var app = builder.Build();
 
 // Configure the HTTP request pipeline
 if (app.Environment.IsDevelopment())
 {
-    app.MapOpenApi();
-    app.MapScalarApiReference(options =>
-    {
-        options.Title = "Expense Tracker API";
-        options.DefaultHttpClient = new(ScalarTarget.CSharp, ScalarClient.HttpClient);
-        options.Theme = ScalarTheme.DeepSpace;
-        options.ShowSidebar = true;
-    });
+    app.UseSwagger();
+    app.UseSwaggerUI();
 }
 
 app.UseHttpsRedirection();
